goldfish_audio: Do not quit QEMU when device fails to initialize
The goldfish audio device is useful but not essential to the Android
emulator. However, in the current implementation, any error during the
initialization of the output voice leads to immediate termination of
QEMU. On the other hand, when such an error occurs, the classic emulator
only prints a warning message and then continues booting.

This issue can be fixed by pretending a successful qdev realization,
i.e. dropping the error_setg() call from goldfish_audio_realize().
Otherwise, once QEMU is notified of the error, an exit() is inevitable.

In addition, move the MMIO initialization code to the beginning of
goldfish_audio_realize() to ensure its execution, because QEMU expects
this qdev to have a MMIO region. If MMIO setup is done in the very end,
any early return statement will result in an assertion error later on.

Change-Id: If0fc0f87553b52456931d9f7e144d91c8a0dac53
Signed-off-by: Yu Ning <yu.ning@intel.com>