UI: Remember the Cellular Network settings
Save the developer's network settings and restore them when the
emulator is launched.

Set up the emulator before the UI, to allow the UI to tell the
device the saved network settings.

Construct the ExtendedWindow early, without showing it. This
ensures that the ExtendedWindow is ready whenever a user opens
it. Previously the window crashed if the user opened it before
it was fully ready.

BUG: b.android.com/197090
BUG: b.android.com/216187

Merged from I1041d2dd480f526ade4e55ada201b3d43c3225a5

Change-Id: I34808cfcf990020121462cbfdae784e3e5f29fad