[UI] Correct emulator's use of port numbers
If the command line indicated the port number(s) for the emulator
to use, the UI and the emulator's interface to adb did not use
the right numbers.

This caused screenshots and file drag/drop to fail. It also
caused incorrect information on the Help : About page.

BUG: b.android.com/216997

Merged from I62955b36aca68efb6c0bdbcc1bd73d5d02bf28df

Change-Id: Ic82440f5c097960d08faf8ccd19449791430f5a7